requirements.txt
streamlit
groq
python-dotenv
pandas
sqlalchemy
langchain
langchain-core
langchain-groq
plotly